+++ src/db/migration/2_create_table_dish.sql
@@ -2,5 +2,5 @@ CREATE TABLE if not exists dish
     (
         id INT PRIMARY KEY,
         name VARCHAR(255),
-        unit_price INT
+        unit_price NUMERIC
     );
